Add example project to demonstrate supporting multiple message types per topic
An end-to-end example project that demonstrates producing and consuming
multiple message types per topic with Schema Registry and Avro / Protobuf.

Not proposing that this is merged or that this is the best long-term approach
for the Confluent.Kafka longer-term. t's just an example of how a team
could extend the current API to support multiple message types per topic.